mesh: Cap sequence overcommit
If application exchanges a large number of messages with a local node
(think: config server, during node initialization), the overcommit
mechanism might save *way* too large sequence number in the storage.

This will cause the node to exhaust sequence numbers if storage fails
between node initialization and a graceful restart. Let's cap the
overcommit, so that this doesn't happen.